Fixed add notification service button
I accidentally committed half of the code for the new URL builder a long time ago, breaking the old method of adding a notification service. Making the URL builder is taking a lot longer than expected, so instead of leaving it broken until I complete the URL builder and then making a release, I've decided to fix it to the old way first, make a release and then include the URL builder in the next release. This means that currently there is a lot of JS code that is not being used or is commented out, because it's part of the URL builder that is coming.